Add files via upload
Walmart is one of the biggest international retail businesses with several outlets around the United States. Walmart plans to analyse sales data from many stores to improve customer satisfaction and optimise store performance over three years (2010 to 2012). The aim is to get a better understanding of sales trends, the effect of holidays, and how sales are correlated with other variables like temperature, fuel prices, the CPI, and unemployment rates.​ R is used for this analysis